// packages/hooks/feed/types.ts
// Generated by https://quicktype.io

export interface OpenGraphType {
  title: string;
  description: string;
  type: string;
  image: Image;
  url: string;
  articlePublishedTime: string;
  articleModifiedTime: string;
  articleSection: string;
  domain: string;
}

export interface Image {
  url: string;
}